+++ g2configmgr/g2configmgr_examples_test.go
@@ -160,7 +160,7 @@ func ExampleG2configmgr_Init() {
 	}
 	moduleName := "Test module name"
 	iniParams := "{}"
-	verboseLogging := 0
+	verboseLogging := int64(0)
 	err := g2configmgr.Init(ctx, moduleName, iniParams, verboseLogging)
 	if err != nil {
 		// This should produce a "senzing-60124002" error.
